+++ bench/d/fib40.d
@@ -1,6 +1,6 @@
 import std.stdio;
 
-double fib(double n)
+int fib(int n)
 {
     if (n < 2)
     {
